Add GenerateSessionID function
This commit introduces the `GenerateSessionID` function, which is responsible for generating a random session ID based on the provided `idType`. The function creates a random string of 12 characters using a set of alphanumeric characters. It then combines the generated string with the `idType` to form the session ID.

This function will be used for session management in the application.