Add GetData function
This commit introduces the `GetData` function, which is used to receive data from a WebSocket connection. The function listens for messages from the WebSocket socket.io server and processes them accordingly. It maintains a buffer of 2MB for incoming data.

- When a message is received, it checks for ping-pong messages ("~h~") and responds appropriately by sending a message back.
- If a message indicates a "timescale_update," it adds the data to the dataset and sends a request for more data.
- If a message indicates that the data transmission is completed with a limit, it exits the loop.

The function returns a list of lists containing the received data.